Application stop

+++ scripts/application_stop.sh
@@ -1,3 +1,7 @@
 #! /bin/bash
 
-killall java
+if [ -z `ps aux | grep java | grep -x "grep java"` ]; then
+  echo "Nothing to do... Skipping"
+else
+  killall java
+fi
